refactor: split func to files
